feat(MultiTenantTestBase.java): create the MultiTenantTestBase with common test functionalities of multi tenant tests
refactor(FederatedMultiTenantTest): move some general functions into the MultiTenantTestBase
